refactor(providers): isolate elevenlabs native fetch adapter to keep providers clean
